Move Setup writing/reading out of ConnectionInner
This moves the last I/O operation out of ConnectionInner.
